Move WatcherEx to a separate file
